+++ src/Observers/ProductMediaObserver.php
@@ -43,11 +43,18 @@ class ProductMediaObserver extends AbstractProductImportObserver
     const ARTEFACT_TYPE = 'media';
 
     /**
-     * The default image label.
+     * The the default image label.
      *
      * @var string
      */
     const DEFAULT_IMAGE_LABEL = 'Image';
+
+    /**
+     * The array with the image information of on row before they'll be converted into artefacts.
+     *
+     * @var array
+     */
+    protected $images = array();
 
     /**
      * The image artefacts that has to be exported.
@@ -57,6 +64,20 @@ class ProductMediaObserver extends AbstractProductImportObserver
     protected $artefacts = array();
 
     /**
+     * The array with names of the images that should be hidden on the product detail page.
+     *
+     * @var array
+     */
+    protected $imagesToHide = array();
+
+    /**
+     * Holds the image values of the main row.
+     *
+     * @var array
+     */
+    protected $mainRow = array();
+
+    /**
      * Process the observer's business logic.
      *
      * @return array The processed row
@@ -64,8 +85,18 @@ class ProductMediaObserver extends AbstractProductImportObserver
     protected function process()
     {
 
-        // initialize the array for the artefacts
+        // reset the values of the parent row, if the SKU changes
+        if ($this->isLastSku($this->getValue(ColumnKeys::SKU)) === false) {
+            $this->mainRow = array();
+        }
+
+        // initialize the array for the artefacts and the hidden images
+        $this->images = array();
         $this->artefacts = array();
+        $this->imagesToHide = array();
+
+        // load the images that has to be hidden on product detail page
+        $this->loadImagesToHide();
 
         // process the images/additional images
         $this->processImages();
@@ -73,6 +104,32 @@ class ProductMediaObserver extends AbstractProductImportObserver
 
         // append the artefacts that has to be exported to the subject
         $this->addArtefacts($this->artefacts);
+    }
+
+    /**
+     * Resolve's the value with the passed colum name from the actual row. If a callback will
+     * be passed, the callback will be invoked with the found value as parameter. If
+     * the value is NULL or empty, the default value will be returned.
+     *
+     * @param string        $name     The name of the column to return the value for
+     * @param mixed|null    $default  The default value, that has to be returned, if the row's value is empty
+     * @param callable|null $callback The callback that has to be invoked on the value, e. g. to format it
+     *
+     * @return mixed|null The, almost formatted, value
+     * @see \TechDivision\Import\Observers\AbstractObserver::getValue()
+     */
+    protected function getImageValue($name, $default = null, callable $callback = null)
+    {
+
+        // query whether or not the a image value is available, return it if yes
+        if ($this->hasValue($name) && $this->isLastSku($this->getValue(ColumnKeys::SKU)) === false) {
+            return $this->mainRow[$name] = $this->getValue($name, $default, $callback);
+        }
+
+        // try to load it from the parent rows
+        if (isset($this->mainRow[$name])) {
+            return $this->mainRow[$name];
+        }
     }
 
     /**
@@ -90,10 +147,18 @@ class ProductMediaObserver extends AbstractProductImportObserver
         // load the parent SKU from the row
         $parentSku = $this->getValue(ColumnKeys::SKU);
 
+        // load the image types
+        $imageTypes = $this->getImageTypes();
+
         // iterate over the available image fields
-        foreach ($this->getImageTypes() as $imageColumnName => $labelColumnName) {
-            // query whether or not, we've a base image
-            if ($image = $this->getValue($imageColumnName)) {
+        foreach ($imageTypes as $imageColumnName => $labelColumnName) {
+            // query whether or not the column contains an image name
+            if ($image = $this->getImageValue($imageColumnName)) {
+                // load the original image path and query whether or not an image with the name already exists
+                if (isset($this->artefacts[$imagePath = $this->getInversedImageMapping($image)])) {
+                    continue;
+                }
+
                 // initialize the label text
                 $labelText = $this->getDefaultImageLabel();
 
@@ -105,25 +170,27 @@ class ProductMediaObserver extends AbstractProductImportObserver
                 // prepare the new base image
                 $artefact = $this->newArtefact(
                     array(
-                        ColumnKeys::STORE_VIEW_CODE    => $storeViewCode,
-                        ColumnKeys::ATTRIBUTE_SET_CODE => $attributeSetCode,
-                        ColumnKeys::IMAGE_PARENT_SKU   => $parentSku,
-                        ColumnKeys::IMAGE_PATH         => $image,
-                        ColumnKeys::IMAGE_PATH_NEW     => $image,
-                        ColumnKeys::IMAGE_LABEL        => $labelText
+                        ColumnKeys::STORE_VIEW_CODE        => $storeViewCode,
+                        ColumnKeys::ATTRIBUTE_SET_CODE     => $attributeSetCode,
+                        ColumnKeys::IMAGE_PARENT_SKU       => $parentSku,
+                        ColumnKeys::IMAGE_PATH             => $imagePath,
+                        ColumnKeys::IMAGE_PATH_NEW         => $image,
+                        ColumnKeys::HIDE_FROM_PRODUCT_PAGE => in_array($image, $this->imagesToHide) ? 1 : 0,
+                        ColumnKeys::IMAGE_LABEL            => $labelText
                     ),
                     array(
-                        ColumnKeys::STORE_VIEW_CODE    => ColumnKeys::STORE_VIEW_CODE,
-                        ColumnKeys::ATTRIBUTE_SET_CODE => ColumnKeys::ATTRIBUTE_SET_CODE,
-                        ColumnKeys::IMAGE_PARENT_SKU   => ColumnKeys::SKU,
-                        ColumnKeys::IMAGE_PATH         => $imageColumnName,
-                        ColumnKeys::IMAGE_PATH_NEW     => $imageColumnName,
-                        ColumnKeys::IMAGE_LABEL        => $labelColumnName
+                        ColumnKeys::STORE_VIEW_CODE        => ColumnKeys::STORE_VIEW_CODE,
+                        ColumnKeys::ATTRIBUTE_SET_CODE     => ColumnKeys::ATTRIBUTE_SET_CODE,
+                        ColumnKeys::IMAGE_PARENT_SKU       => ColumnKeys::SKU,
+                        ColumnKeys::IMAGE_PATH             => $imageColumnName,
+                        ColumnKeys::IMAGE_PATH_NEW         => $imageColumnName,
+                        ColumnKeys::HIDE_FROM_PRODUCT_PAGE => ColumnKeys::HIDE_FROM_PRODUCT_PAGE,
+                        ColumnKeys::IMAGE_LABEL            => $labelColumnName
                     )
                 );
 
                 // append the base image to the artefacts
-                $this->artefacts[] = $artefact;
+                $this->artefacts[$imagePath] = $artefact;
             }
         }
     }
@@ -144,37 +211,61 @@ class ProductMediaObserver extends AbstractProductImportObserver
         $parentSku = $this->getValue(ColumnKeys::SKU);
 
         // query whether or not, we've additional images
-        if ($additionalImages = $this->getValue(ColumnKeys::ADDITIONAL_IMAGES, null, array($this, 'explode'))) {
+        if ($additionalImages = $this->getImageValue(ColumnKeys::ADDITIONAL_IMAGES, null, array($this, 'explode'))) {
             // expand the additional image labels, if available
             $additionalImageLabels = $this->getValue(ColumnKeys::ADDITIONAL_IMAGE_LABELS, array(), array($this, 'explode'));
 
             // initialize the images with the found values
             foreach ($additionalImages as $key => $additionalImage) {
+                // load the original image path and query whether or not an image with the name already exists
+                if (isset($this->artefacts[$imagePath = $this->getInversedImageMapping($additionalImage)])) {
+                    continue;
+                }
+
                 // prepare the additional image
                 $artefact = $this->newArtefact(
                     array(
-                        ColumnKeys::STORE_VIEW_CODE    => $storeViewCode,
-                        ColumnKeys::ATTRIBUTE_SET_CODE => $attributeSetCode,
-                        ColumnKeys::IMAGE_PARENT_SKU   => $parentSku,
-                        ColumnKeys::IMAGE_PATH         => $additionalImage,
-                        ColumnKeys::IMAGE_PATH_NEW     => $additionalImage,
-                        ColumnKeys::IMAGE_LABEL        => isset($additionalImageLabels[$key]) ?
-                                                          $additionalImageLabels[$key] :
-                                                          $this->getDefaultImageLabel()
+                        ColumnKeys::STORE_VIEW_CODE        => $storeViewCode,
+                        ColumnKeys::ATTRIBUTE_SET_CODE     => $attributeSetCode,
+                        ColumnKeys::IMAGE_PARENT_SKU       => $parentSku,
+                        ColumnKeys::IMAGE_PATH             => $imagePath,
+                        ColumnKeys::IMAGE_PATH_NEW         => $additionalImage,
+                        ColumnKeys::HIDE_FROM_PRODUCT_PAGE => in_array($additionalImage, $this->imagesToHide) ? 1 : 0,
+                        ColumnKeys::IMAGE_LABEL            => isset($additionalImageLabels[$key]) ?
+                                                              $additionalImageLabels[$key] :
+                                                              $this->getDefaultImageLabel()
                     ),
                     array(
-                        ColumnKeys::STORE_VIEW_CODE    => ColumnKeys::STORE_VIEW_CODE,
-                        ColumnKeys::ATTRIBUTE_SET_CODE => ColumnKeys::ATTRIBUTE_SET_CODE,
-                        ColumnKeys::IMAGE_PARENT_SKU   => ColumnKeys::SKU,
-                        ColumnKeys::IMAGE_PATH         => ColumnKeys::ADDITIONAL_IMAGES,
-                        ColumnKeys::IMAGE_PATH_NEW     => ColumnKeys::ADDITIONAL_IMAGES,
-                        ColumnKeys::IMAGE_LABEL        => ColumnKeys::ADDITIONAL_IMAGE_LABELS
+                        ColumnKeys::STORE_VIEW_CODE        => ColumnKeys::STORE_VIEW_CODE,
+                        ColumnKeys::ATTRIBUTE_SET_CODE     => ColumnKeys::ATTRIBUTE_SET_CODE,
+                        ColumnKeys::IMAGE_PARENT_SKU       => ColumnKeys::SKU,
+                        ColumnKeys::IMAGE_PATH             => ColumnKeys::ADDITIONAL_IMAGES,
+                        ColumnKeys::IMAGE_PATH_NEW         => ColumnKeys::ADDITIONAL_IMAGES,
+                        ColumnKeys::HIDE_FROM_PRODUCT_PAGE => ColumnKeys::HIDE_FROM_PRODUCT_PAGE,
+                        ColumnKeys::IMAGE_LABEL            => ColumnKeys::ADDITIONAL_IMAGE_LABELS
                     )
                 );
 
                 // append the additional image to the artefacts
-                $this->artefacts[] = $artefact;
+                $this->artefacts[$imagePath] = $artefact;
             }
+        }
+    }
+
+    /**
+     * Load the images that has to be hidden on the product detail page.
+     *
+     * @return void
+     */
+    protected function loadImagesToHide()
+    {
+
+        // load the array with the images that has to be hidden
+        $hideFromProductPage = $this->getValue(ColumnKeys::HIDE_FROM_PRODUCT_PAGE, array(), array($this, 'explode'));
+
+        // map the image names, because probably they have been renamed by the upload functionlity
+        foreach ($hideFromProductPage as $filename) {
+            $this->imagesToHide[] = $this->getImageMapping($filename);
         }
     }
 
@@ -191,11 +282,35 @@ class ProductMediaObserver extends AbstractProductImportObserver
     /**
      * Return's the default image label.
      *
-     * @return string The default image label
+     * @return string|null The default image label
      */
     protected function getDefaultImageLabel()
     {
         return ProductMediaObserver::DEFAULT_IMAGE_LABEL;
+    }
+
+    /**
+     * Returns the mapped filename (which is the new filename).
+     *
+     * @param string $filename The filename to map
+     *
+     * @return string The mapped filename
+     */
+    protected function getImageMapping($filename)
+    {
+        return $this->getSubject()->getImageMapping($filename);
+    }
+
+    /**
+     * Returns the original filename for passed one (which is the new filename).
+     *
+     * @param string $newFilename The new filename to return the original one for
+     *
+     * @return string The original filename
+     */
+    protected function getInversedImageMapping($newFilename)
+    {
+        return $this->getSubject()->getInversedImageMapping($newFilename);
     }
 
     /**
@@ -222,6 +337,6 @@ class ProductMediaObserver extends AbstractProductImportObserver
      */
     protected function addArtefacts(array $artefacts)
     {
-        $this->getSubject()->addArtefacts(ProductMediaObserver::ARTEFACT_TYPE, $artefacts);
+        $this->getSubject()->addArtefacts(ProductMediaObserver::ARTEFACT_TYPE, $artefacts, false);
     }
 }